+++ packages/shopify-app-remix/src/server/authenticate/admin/strategies/auth-code-flow.ts
@@ -304,9 +304,9 @@ export class AuthCodeFlowStrategy<
         });
       }
     } else if (error instanceof GraphqlQueryError) {
-      const context: {[key: string]: string} = {shop};
+      const context: Record<string, string> = {shop};
       if (error.response) {
-        context.response = JSON.stringify(error.response);
+        context.response = JSON.stringify(error.body);
       }
 
       logger.error(